fix: favorite 버그 수정 및 nav-bar 리팩토링
