+++ src/features/favorites/index.ts
@@ -1,5 +1,4 @@
 export { favoritesApi } from './api/favorites.api';
-export { favoriteKeys, useFavoritesCount } from './model/favorites.queries';
-export { useFavoriteMeeting } from './model/use-toggle-favorite';
+export { favoriteKeys, useFavoriteMeeting, useFavoritesCount } from './model/favorites.queries';
 export type { HeartButtonProps } from './ui/heart-button';
 export { HeartButton } from './ui/heart-button';
